requirements.txt
prettytable==3.9.0
pandas==1.5.3
protobuf==3.20.1 
scikit-learn==1.3.2
transformers==4.24.0
beautifulsoup4==4.12.2
sentencepiece==0.1.99
pyparsing==3.1.1
tensorboardx==2.6.2.2
pytorch-warmup==0.1.1
ujson==5.8.0
warcio==1.7.4
chardet==5.2.0
pycld2==0.41
fastapi==0.110.2
uvicorn==0.15.0